Adicionado unidade de teste para "Pessoa" e removido o projeto "WebTests2" e deixado apenas o "WebTests"
