+++ Codigo/Feiragro.sln
@@ -14,6 +14,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Service", "Service\Service.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FeiragroWebTests2", "FeiragroWebTests2\FeiragroWebTests2.csproj", "{47AF92DE-19AE-46C2-A75C-25D9588554D9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FeiragroWebTests", "FeiragroWebTests\FeiragroWebTests.csproj", "{4E41CBE4-1710-41CE-B115-88F8BCEF0A6E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -36,6 +38,10 @@ Global
 		{47AF92DE-19AE-46C2-A75C-25D9588554D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47AF92DE-19AE-46C2-A75C-25D9588554D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{47AF92DE-19AE-46C2-A75C-25D9588554D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4E41CBE4-1710-41CE-B115-88F8BCEF0A6E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4E41CBE4-1710-41CE-B115-88F8BCEF0A6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4E41CBE4-1710-41CE-B115-88F8BCEF0A6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4E41CBE4-1710-41CE-B115-88F8BCEF0A6E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
